Q8_0 dequant+cuBLAS F32 GEMM: 14-16% prefill speedup
- Add dequant_q8_0_to_f16 kernel to MoE cubin (outputs F32 despite name)
- Add fn_dequant_q8_0_to_f16 field + CLLM_LOAD_OPT_FUNC loading
- Add d_f16_scratch buffer (256 MB) for dequantized F32 weights
- Wire into launch_batch_matvec: dequant Q8_0 -> F32 once, cuBLAS F32 GEMM
- Add cuBLAS handle creation + FA2 buffer alloc in generic prefill path
- Add cleanup for d_f16_scratch

VLM prefill: 4858 ms (was 5347 ms, 9% faster)
Text-only pp33: 4277 ms (was 4977 ms, 14% faster)

+++ cuda/llm/moe_gpu_kernels.cu
@@ -730,9 +730,9 @@ extern "C" __global__ void moe_expert_fused_q4k(
     }
 }
 
-/* ---- dequant_q8_0_to_f16: Q8_0 weight matrix -> F16 ---- */
+/* ---- dequant_q8_0_to_f16: Q8_0 weight matrix -> F32 (F16 in name for backwards compat) ---- */
 extern "C" __global__ void dequant_q8_0_to_f16(
-    half *dst,
+    float *dst,
     const unsigned char *mat,
     int rows, int cols)
 {
@@ -745,12 +745,12 @@ extern "C" __global__ void dequant_q8_0_to_f16(
     int bk = bid % nb;
 
     const unsigned char *bp = mat + (size_t)row * rb + (size_t)bk * 36;
-    half *d = dst + (size_t)row * cols + (size_t)bk * 32;
+    float *d = dst + (size_t)row * cols + (size_t)bk * 32;
 
     float scale = __half2float(*(const __half *)bp);
     const signed char *qs = (const signed char *)(bp + 4);
 
     for (int i = 0; i < 32; i++) {
-        d[i] = __float2half(scale * (float)qs[i]);
+        d[i] = scale * (float)qs[i];
     }
 }
